# public/openapi.yaml
openapi: 3.0.3
info:
  title: 同频 B2 API
  version: 1.3.0
  description: 同频 B2 情绪急救与数字陪伴前后端接口。支持 Swagger UI Try it out；可在 Servers 中切换本地或远程环境。注意：本 API 不替代专业医疗或心理服务，AI 输出必须人工确认。
servers:
- url: http://localhost:8080/api/v1
  description: 本地开发环境
- url: https://{remoteHost}/api/v1
  description: 远程环境（修改 remoteHost 为实际域名）
  variables:
    remoteHost:
      default: api.tongpin.example.com
      description: 远程 API 域名，不带 https:// 和路径
...
paths:
...
  /auth/user-session:
    post:
      tags:
      - Auth
      summary: 创建匿名用户会话
      operationId: createUserSession
      description: ''
      security: []
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericObject'
            example:
              age_band: 14-17
              district_id: district_shanghai_a
              client_type: web
              client_version: 1.0.0
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSessionResponse'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 状态冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: 服务不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/volunteer-login:
    post:
      tags:
      - Auth
      summary: 志愿者登录并获取会话 Token
      operationId: volunteerLogin
      description: ''
      security: []
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericObject'
            example:
              account: demo_volunteer
              password: demo-password
              client_type: web
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 状态冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: 服务不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/manager-login:
    post:
      tags:
      - Auth
      summary: 值班负责人或专业督导登录
      operationId: managerLogin
      description: ''
      security: []
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericObject'
            example:
              account: demo_manager
              password: demo-password
              client_type: web
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 状态冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: 服务不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/refresh:
    post:
      tags:
      - Auth
      summary: 刷新访问 Token
      operationId: refreshToken
      description: ''
      security: []
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericObject'
            example:
              refresh_token: refresh_token
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 状态冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: 服务不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/logout:
    post:
      tags:
      - Auth
      summary: 退出登录并撤销当前会话
      operationId: logout
      description: ''
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericObject'
            example:
              session_id: session_01H8YQ
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 状态冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: 服务不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /ai/diary/draft:
    post:
      tags:
      - AI
      summary: 生成日记草稿
      operationId: generateDiaryDraft
      description: ''
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericObject'
            example:
              feeling: low
              free_text: 今天有点累，也不想说话。
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiaryEntry'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 状态冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: 服务不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /ai/chat/suggestion:
    post:
      tags:
      - AI
      summary: 生成小频回复候选
      operationId: generateChatSuggestion
      description: ''
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericObject'
            example:
              case_id: case_01H8YQ
              context_scope:
              - chat_text
              - main_request
              goal: empathetic_listening
              max_tokens: 500
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 状态冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: 服务不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /ai/transfer-summary:
    post:
      tags:
      - AI
      summary: 生成最小交接摘要
      operationId: generateTransferSummary
      description: ''
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericObject'
            example:
              case_id: case_01H8YQ
              summary_scope:
              - main_request
              - safety_facts
              - actions_taken
              - open_questions
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 状态冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: 服务不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /ai/safety-check:
    post:
      tags:
      - AI
      summary: 检查建议是否越界
      operationId: checkAISafety
      description: ''
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericObject'
            example:
              task_type: chat_suggestion
              content: 建议用户立刻拨打120并等待救援。
              context_scope:
              - chat_text
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 状态冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: 服务不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /ai/resource-recommendation:
    post:
      tags:
      - AI
      summary: 检索资源候选
      operationId: recommendResources
      description: ''
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericObject'
            example:
              region: shanghai
              category: mental_hotline
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resource'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 状态冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: 服务不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 输入登录后获取的 access token。匿名用户先调用 /auth/user-session。
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: 写接口幂等键。重复提交同一请求时应返回同一结果。
  schemas:
    Error:
      type: object
      required:
      - error
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
            details:
              type: object
              additionalProperties: true
...
    UserSessionResponse:
      type: object
      required:
      - token
      - user
      - expires_at
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/User'
        expires_at:
          type: string
          format: date-time
    User:
      type: object
      required:
      - id
      - age_band
      - district_id
      properties:
        id:
          type: string
        display_name:
          type: string
        age_band:
          type: string
        district_id:
          type: string
        created_at:
          type: string
          format: date-time
        consents:
          type: object
          additionalProperties: true
...
    DiaryEntry:
      type: object
      properties:
        id:
          type: string
        feeling:
          type: string
        content:
          type: string
        status:
          type: string
        current_version:
          type: integer
        confirmed_at:
          type:
          - string
          - 'null'
          format: date-time
        shared_scope:
          type:
          - string
          - 'null'
...
    Resource:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        category:
          type: string
        region:
          type: string
        phone:
          type: string
        available_time:
          type: string
        official_source:
          type: string
        verified_at:
          type: string
          format: date-time
        status:
          type: string
...
    GenericObject:
      type: object
      additionalProperties: true